@@ -21,21 +21,29 @@
   },
   "dependencies": {
     "element-plus": "^2.4.1",
-    "vue-router": "^4.0.8"
+    "vue-router": "^4.0.8",
+    "lodash-es": "^4.17.21"
   },
   "devDependencies": {
-    "@rushstack/eslint-patch": "^1.3.3",
-    "@vue/eslint-config-prettier": "^8.0.0",
-    "@vue/eslint-config-typescript": "^12.0.0",
+    "@rushstack/eslint-patch": "^1.8.0",
+    "@tsconfig/node20": "^20.1.4",
+    "@types/jsdom": "^21.1.7",
+    "@types/node": "^20.14.5",
+    "@types/lodash-es": "^4.17.12",
+    "@types/qs": "^6.9.15",
+    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-typescript": "^13.0.0",
+    "@vue/tsconfig": "^0.5.1",
     "cross-env": "^7.0.3",
-    "eslint": "^8.49.0",
-    "eslint-plugin-vue": "^9.17.0",
+    "eslint": "^8.57.0",
+    "eslint-plugin-vue": "^9.23.0",
+    "jsdom": "^24.1.0",
+    "npm-run-all2": "^6.2.0",
+    "prettier": "^3.2.5",
     "file-loader": "^6.2.0",
-    "jsdom": "^22.1.0",
-    "npm-run-all2": "^6.1.1",
-    "prettier": "^3.0.3",
-    "typescript": "^5.2.2",
-    "vue-tsc": "^1.8.19"
+    "typescript": "^5.4.5",
+    "vite": "^5.3.1",
+    "vue-tsc": "^2.0.21"
   },
   "pnpm": {
     "overrides": {
@@ -48,7 +56,7 @@
       "@formily/shared": "^2.3.0",
       "@formily/vue": "^2.3.0",
       "element-plus": "^2.4.3",
-      "vue": "^3.3.4"
+      "vue": "^3.4.29"
     }
   }
 }
