@@ -1,4 +1,4 @@
-import { defineComponent, provide, ref } from '@vue/composition-api'
+import { defineComponent, provide, ref } from 'vue'
 import { workerApi } from '@tap/api'
 import { provideI18n, useMessage, I18n } from './hooks'
 import LockedDialog from './components/LockedDialog.vue'
@@ -14,12 +14,12 @@ export default defineComponent({
             alarmSetting: true,
             roleList: true,
             valid_total: true,
-            webhookAlerts: true
+            webhookAlerts: true,
           }
         : {}
 
     provideI18n()
-    provide('checkAgent', async cb => {
+    provide('checkAgent', async (cb) => {
       const Message = useMessage()
       const data = await workerApi.getAvailableAgent()
       if (!data?.result?.length) {
@@ -41,11 +41,11 @@ export default defineComponent({
           <LockedDialog
             visible={showLocked.value}
             on={{
-              'update:visible': val => (showLocked.value = val)
+              'update:visible': (val) => (showLocked.value = val),
             }}
           ></LockedDialog>
         </div>
       )
     }
-  }
+  },
 })
