@@ -50,7 +50,7 @@ export default ({ routes }) => {
     }
     let queryString = ``
     if (process.env.NODE_ENV === 'development') {
-      queryString = `__token=${process.env.VITE_ACCESS_TOKEN}`
+      queryString = `__token=${process.env.VUE_APP_ACCESS_TOKEN}`
     }
     let index = loc.pathname.lastIndexOf('.html')
     let path = loc.pathname
@@ -67,8 +67,8 @@ export default ({ routes }) => {
     window.uetq.push('set', {
       pid: {
         em: window.__USER_INFO__.email,
-        ph: window.__USER_INFO__.telephone
-      }
+        ph: window.__USER_INFO__.telephone,
+      },
     })
 
     /*S 万维广告*/
