@@ -167,12 +167,12 @@ module.exports = defineConfig({
           // 其余配置查看compression-webpack-plugin
         }),
         // ace editor js 输出到 js/ace 目录
-        new webpack.NormalModuleReplacementPlugin(/^file-loader\?esModule=false!\.\/src-noconflict(.*)/, res => {
+        new webpack.NormalModuleReplacementPlugin(/^file-loader\?esModule=false!\.\/src-noconflict(.*)/, (res) => {
           res.request = res.request.replace(
             /^file-loader\?esModule=false!/,
-            'file-loader?esModule=false&outputPath=js/ace!'
+            'file-loader?esModule=false&outputPath=js/ace!',
           )
-        })
+        }),
       )
 
       config['performance'] = {
@@ -335,9 +335,9 @@ const getToken = (userId) => {
 }
 if (process.env.NODE_ENV === 'development') {
   let _userId = process.env.USER_ID || userId
-  process.env.VITE_ACCESS_TOKEN = getToken(_userId)
+  process.env.VUE_APP_ACCESS_TOKEN = getToken(_userId)
 
   console.log('本地用户调试ID: ' + _userId)
-  console.log('本地用户调试Token: ' + process.env.VITE_ACCESS_TOKEN)
+  console.log('本地用户调试Token: ' + process.env.VUE_APP_ACCESS_TOKEN)
   console.log('Proxy server: ' + proxy.target)
 }
