@@ -128,7 +128,7 @@ export default {
       }
     }
     return {
-      isDaas: import.meta.env.VITE_PLATFORM === 'DAAS',
+      isDaas: import.meta.env.VUE_APP_PLATFORM === 'DAAS',
       rules: [],
       id: '',
       commandCallbackFunctionId: '',
@@ -161,7 +161,7 @@ export default {
       // 当前连接是否有共享缓存任务使用
       connectionLogCollectorTaskData: {
         items: [],
-        total: 0
+        total: 0,
       },
       showAgentIpAlert: false,
     }
@@ -185,7 +185,7 @@ export default {
       return `https://docs.tapdata.${
         !this.$store.getters.isDomesticStation || this.$i18n.locale === 'en' ? 'io' : 'net'
       }/prerequisites/allow-access-network`
-    }
+    },
   },
   async created() {
     this.id = this.params.id || ''
@@ -482,7 +482,7 @@ export default {
       // 是否支持共享挖掘
       if (
         !this.lockedFeature.sharedMiningList &&
-        this.pdkOptions.capabilities?.some(t => t.id === 'stream_read_function')
+        this.pdkOptions.capabilities?.some((t) => t.id === 'stream_read_function')
       ) {
         Object.assign(endProperties, {
           shareCdcEnable: {
@@ -684,32 +684,32 @@ export default {
                   value: `{{!$isDaas && $deps[0] ? 'MANUALLY_SPECIFIED_BY_THE_USER' : $self.value}}`,
                   dataSource: `{{!$isDaas && $deps[0] ? [
                     { label: '${this.$t(
-                      'packages_business_connection_form_automatic'
+                      'packages_business_connection_form_automatic',
                     )}', value: 'AUTOMATIC_PLATFORM_ALLOCATION', disabled: true },
                     { label: '${this.$t(
-                      'packages_business_connection_form_manual'
+                      'packages_business_connection_form_manual',
                     )}', value: 'MANUALLY_SPECIFIED_BY_THE_USER' }
                   ] : !$isDaas ? [
                     { label: '${this.$t(
-                      'packages_business_connection_form_automatic'
+                      'packages_business_connection_form_automatic',
                     )}', value: 'AUTOMATIC_PLATFORM_ALLOCATION' },
                     { label: '${this.$t(
-                      'packages_business_connection_form_manual'
+                      'packages_business_connection_form_manual',
                     )}', value: 'MANUALLY_SPECIFIED_BY_THE_USER' }
                   ] : [
                     { label: '${this.$t(
-                      'packages_business_connection_form_automatic'
+                      'packages_business_connection_form_automatic',
                     )}', value: 'AUTOMATIC_PLATFORM_ALLOCATION' },
                     { label: '${this.$t(
-                      'packages_business_connection_form_manual'
+                      'packages_business_connection_form_manual',
                     )}', value: 'MANUALLY_SPECIFIED_BY_THE_USER' },
                     {
                       label: '${this.$t('packages_business_connection_form_group')}',
                       value: 'MANUALLY_SPECIFIED_BY_THE_USER_AGENT_GROUP'
                     }
-                  ]}}`
-                }
-              }
+                  ]}}`,
+                },
+              },
             },
             {
               target: '__TAPDATA.accessNodeProcessId',
@@ -732,42 +732,42 @@ export default {
               fulfill: {
                 state: {
                   visible:
-                    "{{['MANUALLY_SPECIFIED_BY_THE_USER', 'MANUALLY_SPECIFIED_BY_THE_USER_AGENT_GROUP'].includes($deps[0])}}"
-                }
-              }
+                    "{{['MANUALLY_SPECIFIED_BY_THE_USER', 'MANUALLY_SPECIFIED_BY_THE_USER_AGENT_GROUP'].includes($deps[0])}}",
+                },
+              },
             },
-            '{{useAsyncDataSource(loadAccessNode, "dataSource", {value: $self.value})}}'
-          ]
+            '{{useAsyncDataSource(loadAccessNode, "dataSource", {value: $self.value})}}',
+          ],
         },
         agentWrap: {
           type: 'void',
           'x-component': 'Space',
           'x-component-props': {
-            class: 'w-100 align-items-start'
+            class: 'w-100 align-items-start',
           },
           'x-reactions': {
             dependencies: ['.accessNodeType'],
             fulfill: {
               state: {
                 visible:
-                  "{{['MANUALLY_SPECIFIED_BY_THE_USER', 'MANUALLY_SPECIFIED_BY_THE_USER_AGENT_GROUP'].includes($deps[0])}}"
-              }
-            }
+                  "{{['MANUALLY_SPECIFIED_BY_THE_USER', 'MANUALLY_SPECIFIED_BY_THE_USER_AGENT_GROUP'].includes($deps[0])}}",
+              },
+            },
           },
           properties: {
             accessNodeProcessId: {
               type: 'string',
               description: `{{$values.__TAPDATA.shareCdcEnable ? '${this.$t(
-                'packages_business_agent_select_not_found_for_rocksdb'
+                'packages_business_agent_select_not_found_for_rocksdb',
               )}' : ''}}`,
               'x-decorator': 'FormItem',
               'x-decorator-props': {
                 colon: false,
-                class: 'flex-1'
+                class: 'flex-1',
               },
               'x-component': 'Select',
               'x-component-props': {
-                onChange: `{{ () => $self.setSelfErrors('') }}`
+                onChange: `{{ () => $self.setSelfErrors('') }}`,
               },
               'x-reactions': [
                 // '{{useAsyncDataSource(loadAccessNode, "dataSource", {value: $self.value})}}',
@@ -777,8 +777,8 @@ export default {
                   fulfill: {
                     state: {
                       title: `{{'MANUALLY_SPECIFIED_BY_THE_USER_AGENT_GROUP' === $deps[0] ? '${i18n.t(
-                        'packages_business_choose_agent_group'
-                      )}': '${i18n.t('packages_business_choose_agent')}'}}`
+                        'packages_business_choose_agent_group',
+                      )}': '${i18n.t('packages_business_choose_agent')}'}}`,
                     },
                     run: `
                 console.log('$deps[1]', $deps)
@@ -792,9 +792,9 @@ export default {
                     $self.setSelfErrors('${this.$t('packages_business_agent_select_not_found')}')
                   }
                 }
-              }`
-                  }
-                }
+              }`,
+                  },
+                },
               ],
               // 校验下拉数据判断是否存在已选的agent
               'x-validator': `{{(value, rule, ctx)=> {
@@ -814,7 +814,7 @@ export default {
                 return '${this.$t('packages_business_agent_select_not_found')}'
               }
             }
-          }}}`
+          }}}`,
             },
             priorityProcessId: {
               title: i18n.t('packages_business_priorityProcessId'),
@@ -822,14 +822,14 @@ export default {
               default: '',
               'x-decorator': 'FormItem',
               'x-decorator-props': {
-                class: 'flex-1'
+                class: 'flex-1',
               },
               'x-component': 'Select',
               'x-reactions': {
                 dependencies: ['.accessNodeType', '.accessNodeOption#dataSource', '.accessNodeProcessId'],
                 fulfill: {
                   state: {
-                    visible: "{{'MANUALLY_SPECIFIED_BY_THE_USER_AGENT_GROUP' === $deps[0]}}"
+                    visible: "{{'MANUALLY_SPECIFIED_BY_THE_USER_AGENT_GROUP' === $deps[0]}}",
                   },
                   run: `
                     let children = []
@@ -848,11 +848,11 @@ export default {
                     if ($self.value && !children.find(item => item.value === $self.value)) {
                       $self.value = ''
                     }
-                  `
-                }
-              }
-            }
-          }
+                  `,
+                },
+              },
+            },
+          },
         },
         schemaUpdateHour: {
           type: 'string',
@@ -928,7 +928,7 @@ export default {
         })
         endProperties.accessNodeType.enum.push({
           label: this.$t('packages_business_connection_form_group'),
-          value: 'MANUALLY_SPECIFIED_BY_THE_USER_AGENT_GROUP'
+          value: 'MANUALLY_SPECIFIED_BY_THE_USER_AGENT_GROUP',
         })
       }
 
@@ -950,9 +950,9 @@ export default {
                   target: process.env.VUE_APP_HIDE_CONNECTOR_SCHEMA,
                   fulfill: {
                     state: {
-                      display: 'hidden'
-                    }
-                  }
+                      display: 'hidden',
+                    },
+                  },
                 }
               : undefined,
             properties: {
@@ -967,8 +967,8 @@ export default {
                     'x-component': 'Input',
                     'x-validator': {
                       pattern: /^([\u4e00-\u9fa5]|[A-Za-z])([a-zA-Z0-9_\s-.]|[\u4e00-\u9fa5])*$/,
-                      message: i18n.t('packages_business_connections_databaseform_mingchengguizezhong')
-                    }
+                      message: i18n.t('packages_business_connections_databaseform_mingchengguizezhong'),
+                    },
                   },
                   connection_type: {
                     type: 'string',
@@ -1104,9 +1104,9 @@ export default {
                         'x-decorator': 'FormItem',
                         'x-component': 'TextFileReader',
                         'x-component-props': {
-                          base64: true
+                          base64: true,
                         },
-                        fileNameField: 'sslCAFile'
+                        fileNameField: 'sslCAFile',
                       },
                       sslCert: {
                         // 客户端证书文件
@@ -1115,9 +1115,9 @@ export default {
                         'x-decorator': 'FormItem',
                         'x-component': 'TextFileReader',
                         'x-component-props': {
-                          base64: true
+                          base64: true,
                         },
-                        fileNameField: 'sslCertFile'
+                        fileNameField: 'sslCertFile',
                       },
                       sslKey: {
                         // 客户端密钥文件
@@ -1126,9 +1126,9 @@ export default {
                         'x-decorator': 'FormItem',
                         'x-component': 'TextFileReader',
                         'x-component-props': {
-                          base64: true
+                          base64: true,
                         },
-                        fileNameField: 'sslKeyFile'
+                        fileNameField: 'sslKeyFile',
                       },
                       sslKeyPassword: {
                         // 客户端密钥密码
@@ -1247,32 +1247,32 @@ export default {
         loadAccessNode: async (fieldName, others = {}) => {
           const data = await clusterApi.findAccessNodeInfo()
 
-          const mapNode = item => ({
+          const mapNode = (item) => ({
             value: item.processId,
             label: `${item.hostName}（${
               item.status === 'running' ? i18n.t('public_status_running') : i18n.t('public_agent_status_offline')
             }）`,
             disabled: item.status !== 'running',
-            accessNodeType: item.accessNodeType
+            accessNodeType: item.accessNodeType,
           })
 
           return (
             data
               ?.filter(
-                t =>
+                (t) =>
                   t.status === 'running' ||
                   t.accessNodeType === 'MANUALLY_SPECIFIED_BY_THE_USER_AGENT_GROUP' ||
-                  t.processId === others.value
+                  t.processId === others.value,
               )
-              ?.map(item => {
+              ?.map((item) => {
                 if (item.accessNodeType === 'MANUALLY_SPECIFIED_BY_THE_USER_AGENT_GROUP') {
                   return {
                     value: item.processId,
                     label: `${item.accessNodeName}（${i18n.t('public_status_running')}：${
-                      item.accessNodes?.filter(ii => ii.status === 'running').length || 0
+                      item.accessNodes?.filter((ii) => ii.status === 'running').length || 0
                     }）`,
                     accessNodeType: item.accessNodeType,
-                    children: item.accessNodes?.map(mapNode) || []
+                    children: item.accessNodes?.map(mapNode) || [],
                   }
                 }
                 return mapNode(item)
@@ -1324,7 +1324,7 @@ export default {
             expireSeconds: 100000000,
           }
           proxyApi.subscribe(filter).then((data) => {
-            const isDaas = import.meta.env.VITE_PLATFORM === 'DAAS'
+            const isDaas = import.meta.env.VUE_APP_PLATFORM === 'DAAS'
             const p = location.origin + location.pathname
             let str = `${p}${isDaas ? '' : 'tm/'}api/proxy/callback/${data.token}`
             if (/^\/\w+/.test(data.token)) {
@@ -1348,7 +1348,7 @@ export default {
             connectionConfig: isEmpty(formValues) ? this.model?.config || {} : getValues,
             ...others,
             subscribeIds,
-            type: 'connection'
+            type: 'connection',
           }
           proxyApi.command(params).then((data) => {
             const setValue = data.setValue
@@ -1364,16 +1364,16 @@ export default {
             let filter = {
               where: {},
               limit: 1000,
-              skip: 0
+              skip: 0,
             }
             if (id) {
               const ext = await externalStorageApi.get(id)
               filter.where.type = ext?.type
             }
             const { items = [] } = await externalStorageApi.list({
-              filter: JSON.stringify(filter)
+              filter: JSON.stringify(filter),
             })
-            return items.map(item => {
+            return items.map((item) => {
               return {
                 type: item.type,
                 label: item.name,
@@ -1427,7 +1427,7 @@ export default {
           submitForm(params?.target, data)
         },
         shareCDCExternalStorageIdOnChange: (val, $form) => {
-          $form.setFieldState('__TAPDATA.shareCDCExternalStorageIdTips', state => {
+          $form.setFieldState('__TAPDATA.shareCDCExternalStorageIdTips', (state) => {
             state.display =
               this.connectionLogCollectorTaskData.total && val !== this.model.shareCDCExternalStorageId
                 ? 'visible'
@@ -1443,7 +1443,7 @@ export default {
         handleJsDebug: (path = []) => {
           const properties = this.schemaData?.properties || {}
           let fieldObj = {}
-          path.forEach(p => {
+          path.forEach((p) => {
             const { key, data } = this.getOptionByPath(properties, p)
             fieldObj[key] = data
           })
@@ -1451,7 +1451,7 @@ export default {
           this.showJsDebug = true
         },
         handleGetGenerateRefreshToken: ($index, $record, items, others) => {
-          if (items.filter((t, i) => i !== $index).some(t => t.supplierKey === $record.supplierKey)) {
+          if (items.filter((t, i) => i !== $index).some((t) => t.supplierKey === $record.supplierKey)) {
             return this.$message.error(this.$t('packages_form_message_exists_name'))
           }
           const params = Object.assign(
@@ -1459,9 +1459,9 @@ export default {
               supplierKey: $record.supplierKey,
               randomId: $record.randomId,
               subscribeId: `source#${this.model?.id || this.commandCallbackFunctionId}`,
-              service: 'engine'
+              service: 'engine',
             },
-            others
+            others,
           )
           proxyApi.generateRefreshToken(params).then((data = {}) => {
             const isDaas = process.env.VUE_APP_PLATFORM === 'DAAS'
@@ -1479,7 +1479,7 @@ export default {
         getHost: async () => {
           const data = await proxyApi.host()
           return data?.host
-        }
+        },
       }
       this.schemaData = result
       this.loadingFrom = false
