@@ -44,7 +44,7 @@ export default {
   packages_business_connection_form_rename: '改名',
   packages_business_connection_form_database_owner_tip: '逗号分割的表达式列表，使用 * 代表任意长度任意字符',
   packages_business_connection_form_source_and_target_tip: `此数据连接在 ${
-    import.meta.env.VITE_PAGE_TITLE
+    import.meta.env.VUE_APP_PAGE_TITLE
   } 中能同时作为源和目标使用`,
   packages_business_connection_form_source_tip: '此数据连接在Tapdata 中只能作为源使用，不能作用为目标',
   packages_business_connection_form_target_tip: '此数据连接在Tapdata 中只能作为目标使用，不能作用为源',
@@ -1258,5 +1258,5 @@ export default {
   packages_business_shared_cache_enforceShareCdc: '当共享挖掘不可用(缓存启动时)',
   packages_business_shared_cache_enforceShareCdc_true: '缓存直接报错停止',
   packages_business_not_support_validation: '{connection} 不支持{method}',
-  packages_business_download_details: '下载详情'
+  packages_business_download_details: '下载详情',
 }
