@@ -14,207 +14,212 @@
         <span> {{ $t('public_button_create') }}</span>
       </ElButton>
     </template>
-    <section :class="{ paddingLeft0: isDaas }" class="connection-list-wrap rounded-lg">
-      <TablePage
-        ref="table"
-        row-key="id"
-        :classify="
-          isDaas
-            ? {
-                authority: 'datasource_catalog_management',
-                types: ['database'],
-                viewPage: 'connections',
-                title: $t('public_tags')
-              }
-            : null
-        "
-        :remoteMethod="getData"
-        @selection-change="handleSelectionChange"
-        @classify-submit="handleOperationClassify"
-        @sort-change="handleSortTable"
-      >
-        <template #search>
-          <FilterBar v-model="searchParams" :items="filterItems" @fetch="table.fetch(1)">
-            <template #connectionType>
-              <ElRadioGroup v-model="searchParams.databaseModel" size="mini" @change="table.fetch(1)">
-                <ElRadioButton label="">{{ $t('public_all') }}</ElRadioButton>
-                <ElRadioButton label="source">{{ $t('public_connection_type_source') }}</ElRadioButton>
-                <ElRadioButton label="target">{{ $t('public_connection_type_target') }}</ElRadioButton>
-              </ElRadioGroup>
-            </template>
-          </FilterBar>
-        </template>
-        <template v-if="isDaas" #multipleSelectionActions>
-          <ElButton @click="handlePermissionsSettings"
-            >{{ $t('packages_business_permissionse_settings_create_quanxianshezhi') }}
-          </ElButton>
-          <ElButton
-            v-readonlybtn="'datasource_category_application'"
-            class="btn"
-            @click="$refs.table.showClassify(handleSelectTag())"
-          >
-            <i class="iconfont icon-biaoqian back-btn-icon"></i>
-            <span> {{ $t('public_button_bulk_tag') }}</span>
-          </ElButton>
-        </template>
-        <ElTableColumn
-          v-if="isDaas"
-          type="selection"
-          width="38"
-          align="center"
-          :reserve-selection="true"
-        ></ElTableColumn>
-        <ElTableColumn show-overflow-tooltip prop="name" min-width="250" :label="$t('public_connection_name')">
-          <template #default="{ row }">
-            <div class="connection-name flex flex-wrap gap-1">
-              <div class="flex align-center gap-1 overflow-hidden">
-                <img class="connection-img" :src="getConnectionIcon(row.pdkHash)" alt="" />
-                <ElLink class="ellipsis block lh-base" type="primary" @click.stop="preview(row)">
-                  {{ row.name }}
-                </ElLink>
-              </div>
-              <div v-if="row.listtags" class="justify-content-start ellipsis flex flex-wrap align-center gap-1">
-                <span class="tag ellipsis" v-for="(item, i) in row.listtags" :key="i" :title="item.value">
-                  {{ item.value }}
-                </span>
-              </div>
+
+    <TablePage
+      ref="table"
+      row-key="id"
+      :classify="
+        isDaas
+          ? {
+              authority: 'datasource_catalog_management',
+              types: ['database'],
+              viewPage: 'connections',
+              title: $t('public_tags'),
+            }
+          : null
+      "
+      :remoteMethod="getData"
+      @selection-change="handleSelectionChange"
+      @classify-submit="handleOperationClassify"
+      @sort-change="handleSortTable"
+    >
+      <template #search>
+        <FilterBar v-model:value="searchParams" :items="filterItems" @fetch="table.fetch(1)">
+          <template #connectionType>
+            <ElRadioGroup v-model="searchParams.databaseModel" size="mini" @change="table.fetch(1)">
+              <ElRadioButton label="">{{ $t('public_all') }}</ElRadioButton>
+              <ElRadioButton label="source">{{ $t('public_connection_type_source') }}</ElRadioButton>
+              <ElRadioButton label="target">{{ $t('public_connection_type_target') }}</ElRadioButton>
+            </ElRadioGroup>
+          </template>
+        </FilterBar>
+      </template>
+      <template v-if="isDaas" #multipleSelectionActions>
+        <ElButton @click="handlePermissionsSettings"
+          >{{ $t('packages_business_permissionse_settings_create_quanxianshezhi') }}
+        </ElButton>
+        <ElButton
+          v-readonlybtn="'datasource_category_application'"
+          class="btn"
+          @click="$refs.table.showClassify(handleSelectTag())"
+        >
+          <i class="iconfont icon-biaoqian back-btn-icon"></i>
+          <span> {{ $t('public_button_bulk_tag') }}</span>
+        </ElButton>
+      </template>
+      <ElTableColumn v-if="isDaas" type="selection" width="38" align="center" :reserve-selection="true"></ElTableColumn>
+      <ElTableColumn show-overflow-tooltip prop="name" min-width="250" :label="$t('public_connection_name')">
+        <template #default="{ row }">
+          <div class="connection-name flex flex-wrap gap-1">
+            <div class="flex align-center gap-1 overflow-hidden">
+              <img class="connection-img" :src="getConnectionIcon(row.pdkHash)" alt="" />
+              <ElLink class="ellipsis block lh-base" type="primary" @click.stop="preview(row)">
+                {{ row.name }}
+              </ElLink>
             </div>
-          </template>
-        </ElTableColumn>
-        <ElTableColumn show-overflow-tooltip :label="$t('public_connection_information')" min-width="160">
-          <template slot-scope="scope">
-            {{ scope.row.connectionUrl }}
-          </template>
-        </ElTableColumn>
-        <ElTableColumn prop="status" :label="$t('packages_business_connection_dataBaseStatus')" min-width="100">
-          <template #default="{ row }">
-            <div>
-              <span :class="['status-connection-' + row.status, 'status-block']">
-                {{ getStatus(row.status) }}
+            <div v-if="row.listtags" class="justify-content-start ellipsis flex flex-wrap align-center gap-1">
+              <span class="tag ellipsis" v-for="(item, i) in row.listtags" :key="i" :title="item.value">
+                {{ item.value }}
               </span>
             </div>
-          </template>
-        </ElTableColumn>
-        <ElTableColumn width="125" prop="connection_type" min-width="135" :label="$t('public_connection_type')">
-          <template slot-scope="scope">
-            {{ getType(scope.row.connection_type) }}
-          </template>
-        </ElTableColumn>
-        <ElTableColumn min-width="125">
-          <template #header>
-            <div class="flex align-center">
-              <span>{{ $t('public_connection_schema_status') }}</span>
-              <ElTooltip class="ml-2" placement="top" :content="$t('public_connection_schema_status_tip')">
-                <VIcon class="color-primary" size="14">info</VIcon>
-              </ElTooltip>
-            </div>
-          </template>
-          <template v-slot="scope">
-            <div v-if="isFileSource(scope.row)">-</div>
-            <SchemaProgress :data="scope.row"></SchemaProgress>
-          </template>
-        </ElTableColumn>
-        <ElTableColumn
-          prop="loadSchemaTime"
-          sortable="loadSchemaTime"
-          min-width="180"
-          :label="$t('public_connection_table_structure_update_time')"
-        >
-          <template slot-scope="scope">
-            {{ scope.row.loadSchemaTimeLabel }}
-          </template>
-        </ElTableColumn>
-        <ElTableColumn fixed="right" width="320" :label="$t('public_operation')">
-          <div v-if="isDaas" slot="header" class="flex align-center">
+          </div>
+        </template>
+      </ElTableColumn>
+      <ElTableColumn show-overflow-tooltip :label="$t('public_connection_information')" min-width="160">
+        <template v-slot="scope">
+          {{ scope.row.connectionUrl }}
+        </template>
+      </ElTableColumn>
+      <ElTableColumn prop="status" :label="$t('packages_business_connection_dataBaseStatus')" min-width="100">
+        <template #default="{ row }">
+          <div>
+            <span :class="['status-connection-' + row.status, 'status-block']">
+              {{ getStatus(row.status) }}
+            </span>
+          </div>
+        </template>
+      </ElTableColumn>
+      <ElTableColumn width="125" prop="connection_type" min-width="135" :label="$t('public_connection_type')">
+        <template v-slot="scope">
+          {{ getType(scope.row.connection_type) }}
+        </template>
+      </ElTableColumn>
+      <ElTableColumn min-width="125">
+        <template #header>
+          <div class="flex align-center">
+            <span>{{ $t('public_connection_schema_status') }}</span>
+            <ElTooltip class="ml-2" placement="top" :content="$t('public_connection_schema_status_tip')">
+              <VIcon class="color-primary" size="14">info</VIcon>
+            </ElTooltip>
+          </div>
+        </template>
+        <template v-slot="scope">
+          <div v-if="isFileSource(scope.row)">-</div>
+          <SchemaProgress :data="scope.row"></SchemaProgress>
+        </template>
+      </ElTableColumn>
+      <ElTableColumn
+        prop="loadSchemaTime"
+        sortable="loadSchemaTime"
+        min-width="180"
+        :label="$t('public_connection_table_structure_update_time')"
+      >
+        <template v-slot="scope">
+          {{ scope.row.loadSchemaTimeLabel }}
+        </template>
+      </ElTableColumn>
+      <ElTableColumn fixed="right" width="320" :label="$t('public_operation')">
+        <template v-slot:header>
+          <div v-if="isDaas" class="flex align-center">
             <span>{{ $t('public_operation_available') }}</span>
             <ElTooltip class="ml-2" placement="top" :content="$t('packages_business_connections_list_wuquanxiandecao')">
               <VIcon class="color-primary" size="14">info</VIcon>
             </ElTooltip>
           </div>
-          <template slot-scope="scope">
-            <ElButton data-testid="test-connection" type="text" @click="testConnection(scope.row)"
-              >{{ $t('public_connection_button_test') }}
-            </ElButton>
-            <ElDivider direction="vertical"></ElDivider>
-            <ElTooltip
-              :disabled="!isFileSource(scope.row)"
-              :content="$t('packages_business_connections_list_wenjianleixingde')"
-              placement="top"
-            >
-              <span>
-                <ElButton
-                  type="text"
-                  data-testid="load-schema"
-                  :disabled="isFileSource(scope.row) || scope.row.disabledLoadSchema"
-                  @click="handleLoadSchema(scope.row)"
-                  >{{ $t('public_connection_button_load_schema') }}
-                </ElButton>
-              </span>
-            </ElTooltip>
-            <ElDivider direction="vertical"></ElDivider>
-            <ElButton
-              v-if="havePermission(scope.row.permissionActions, 'Edit')"
-              v-readonlybtn="'datasource_edition'"
-              type="text"
-              data-testid="edit-connection"
-              :disabled="
-                $disabledByPermission('datasource_edition_all_data', scope.row.user_id) ||
-                $disabledReadonlyUserBtn() ||
-                scope.row.agentType === 'Cloud'
-              "
-              @click="edit(scope.row.id, scope.row)"
-              >{{ $t('public_button_edit') }}
-            </ElButton>
-            <ElDivider
-              v-if="havePermission(scope.row.permissionActions, 'Edit')"
-              direction="vertical"
-              v-readonlybtn="'datasource_edition'"
-            ></ElDivider>
-            <ElButton
-              v-if="buttonShowMap.copy"
-              v-readonlybtn="'datasource_creation'"
-              type="text"
-              data-testid="copy-connection"
-              :loading="scope.row.copyLoading"
-              :disabled="$disabledReadonlyUserBtn() || scope.row.agentType === 'Cloud'"
-              @click="copy(scope.row)"
-              >{{ $t('public_button_copy') }}
-            </ElButton>
-            <ElDivider v-if="buttonShowMap.copy" direction="vertical" v-readonlybtn="'datasource_creation'"></ElDivider>
-            <ElButton
-              v-if="havePermission(scope.row.permissionActions, 'Delete')"
-              v-readonlybtn="'datasource_delete'"
-              type="text"
-              data-testid="delete-connection"
-              :disabled="
-                $disabledByPermission('datasource_delete_all_data', scope.row.user_id) ||
-                $disabledReadonlyUserBtn() ||
-                scope.row.agentType === 'Cloud'
-              "
-              @click="remove(scope.row)"
-              >{{ $t('public_button_delete') }}
-            </ElButton>
-          </template>
-        </ElTableColumn>
-      </TablePage>
-      <Preview ref="preview" @test="testConnection"></Preview>
-      <!--    <DatabaseTypeDialog
-      :dialogVisible="dialogDatabaseTypeVisible"
-      @dialogVisible="handleDialogDatabaseTypeVisible"
-      @databaseType="handleDatabaseType"
-    ></DatabaseTypeDialog>-->
-      <SceneDialog
-        ref="dialog"
-        :visible.sync="dialogDatabaseTypeVisible"
-        selector-type="source_and_target"
-        v-bind="connectionDialogProps"
-        @selected="handleDatabaseType"
-        @success="table.fetch(1)"
-      ></SceneDialog>
-      <Test ref="test" :visible.sync="dialogTestVisible" :formData="testData" @returnTestData="returnTestData"></Test>
-      <UsedTaskDialog v-model="connectionTaskDialog" :data="connectionTaskData"></UsedTaskDialog>
-      <PermissionseSettingsCreate ref="permissionseSettingsCreate"></PermissionseSettingsCreate>
-    </section>
+        </template>
+        <template v-slot="scope">
+          <ElButton data-testid="test-connection" text type="primary" @click="testConnection(scope.row)"
+            >{{ $t('public_connection_button_test') }}
+          </ElButton>
+          <ElDivider class="mx-1" direction="vertical"></ElDivider>
+          <ElTooltip
+            :disabled="!isFileSource(scope.row)"
+            :content="$t('packages_business_connections_list_wenjianleixingde')"
+            placement="top"
+          >
+            <span>
+              <ElButton
+                text
+                type="primary"
+                data-testid="load-schema"
+                :disabled="isFileSource(scope.row) || scope.row.disabledLoadSchema"
+                @click="handleLoadSchema(scope.row)"
+                >{{ $t('public_connection_button_load_schema') }}
+              </ElButton>
+            </span>
+          </ElTooltip>
+          <ElDivider class="mx-1" direction="vertical"></ElDivider>
+          <ElButton
+            v-if="havePermission(scope.row.permissionActions, 'Edit')"
+            v-readonlybtn="'datasource_edition'"
+            text
+            type="primary"
+            data-testid="edit-connection"
+            :disabled="
+              $disabledByPermission('datasource_edition_all_data', scope.row.user_id) ||
+              $disabledReadonlyUserBtn() ||
+              scope.row.agentType === 'Cloud'
+            "
+            @click="edit(scope.row.id, scope.row)"
+            >{{ $t('public_button_edit') }}
+          </ElButton>
+          <ElDivider
+            class="mx-1"
+            v-if="havePermission(scope.row.permissionActions, 'Edit')"
+            direction="vertical"
+            v-readonlybtn="'datasource_edition'"
+          ></ElDivider>
+          <ElButton
+            v-if="buttonShowMap.copy"
+            v-readonlybtn="'datasource_creation'"
+            text
+            type="primary"
+            data-testid="copy-connection"
+            :loading="scope.row.copyLoading"
+            :disabled="$disabledReadonlyUserBtn() || scope.row.agentType === 'Cloud'"
+            @click="copy(scope.row)"
+            >{{ $t('public_button_copy') }}
+          </ElButton>
+          <ElDivider
+            class="mx-1"
+            v-if="buttonShowMap.copy"
+            direction="vertical"
+            v-readonlybtn="'datasource_creation'"
+          ></ElDivider>
+          <ElButton
+            v-if="havePermission(scope.row.permissionActions, 'Delete')"
+            v-readonlybtn="'datasource_delete'"
+            text
+            type="primary"
+            data-testid="delete-connection"
+            :disabled="
+              $disabledByPermission('datasource_delete_all_data', scope.row.user_id) ||
+              $disabledReadonlyUserBtn() ||
+              scope.row.agentType === 'Cloud'
+            "
+            @click="remove(scope.row)"
+            >{{ $t('public_button_delete') }}
+          </ElButton>
+        </template>
+      </ElTableColumn>
+    </TablePage>
+    <Preview ref="preview" @test="testConnection"></Preview>
+    <!--    <DatabaseTypeDialog
+    :dialogVisible="dialogDatabaseTypeVisible"
+    @dialogVisible="handleDialogDatabaseTypeVisible"
+    @databaseType="handleDatabaseType"
+  ></DatabaseTypeDialog>-->
+    <SceneDialog
+      ref="dialog"
+      v-model:visible="dialogDatabaseTypeVisible"
+      selector-type="source_and_target"
+      v-bind="connectionDialogProps"
+      @selected="handleDatabaseType"
+      @success="table.fetch(1)"
+    ></SceneDialog>
+    <Test ref="test" v-model:visible="dialogTestVisible" :formData="testData" @returnTestData="returnTestData"></Test>
+    <UsedTaskDialog v-model="connectionTaskDialog" :data="connectionTaskData"></UsedTaskDialog>
+    <PermissionseSettingsCreate ref="permissionseSettingsCreate"></PermissionseSettingsCreate>
   </PageContainer>
 </template>
 
@@ -266,7 +271,7 @@ export default {
     }
 
     return {
-      isDaas: import.meta.env.VITE_PLATFORM === 'DAAS',
+      isDaas: import.meta.env.VUE_APP_PLATFORM === 'DAAS',
 
       filterItems: [],
       user_id: Cookie.get('user_id'),
@@ -330,7 +335,7 @@ export default {
       },
       connectionTaskDialog: false,
 
-      connectionDialogProps
+      connectionDialogProps,
     }
   },
   computed: {
@@ -776,6 +781,8 @@ export default {
       } else {
         this.$message.error(this.$t('public_connection_button_test') + this.$t('public_status_invalid'), false)
       }
+      f
+
       this.buried('connectionTest', '', {
         result: status === 'ready',
       })
@@ -795,13 +802,13 @@ export default {
           key: 'status',
           type: 'select-inner',
           items: this.databaseStatusOptions,
-          selectedWidth: '200px',
         },
         {
           label: this.$t('packages_business_connection_list_form_database_type'),
           key: 'databaseType',
           type: 'select-inner',
-          dropdownWidth: '250px',
+          width: '250px',
+          // dropdownWidth: '250px',
           filterable: true,
           items: async () => {
             let data = await databaseTypesApi.get()
@@ -876,80 +883,74 @@ export default {
   height: 100%;
   overflow: hidden;
   background: #fff;
+}
 
-  :deep(.el-select-dropdown__item) {
-    span {
-      font-size: $fontBaseTitle;
-    }
+.connection-name {
+  display: flex;
+  align-items: center;
+}
+
+.tag {
+  padding: 0 4px;
+  font-style: normal;
+  font-weight: 400;
+  font-size: 12px;
+  line-height: 20px;
+  color: map-get($color, tag);
+  border: 1px solid map-get($bgColor, tag);
+  border-radius: 4px;
+}
+
+.connection-img {
+  width: 18px;
+  height: 18px;
+}
+
+.btn-text {
+  // color: map-get($color, primary);
+  font-size: 12px;
+  padding-right: 5px;
+}
+
+.error {
+  color: #f56c6c;
+}
+
+.success {
+  color: #67c23a;
+}
+
+.warning {
+  color: #e6a23c;
+}
+
+.connections-status__icon {
+  font-size: 14px;
+}
+
+.search-bar {
+  display: flex;
+
+  .item {
+    margin-right: 10px;
   }
+}
 
-  .connection-name {
-    display: flex;
-    align-items: center;
-  }
+.btn + .btn {
+  margin-left: 5px;
+}
 
-  .tag {
-    padding: 0 4px;
-    font-style: normal;
-    font-weight: 400;
+.btn {
+  i.iconfont {
     font-size: 12px;
-    line-height: 20px;
-    color: map-get($color, tag);
-    border: 1px solid map-get($bgColor, tag);
-    border-radius: 4px;
   }
 
-  .connection-img {
-    width: 18px;
-    height: 18px;
-  }
-
-  .btn-text {
-    // color: map-get($color, primary);
-    font-size: 12px;
-    padding-right: 5px;
-  }
-
-  .error {
-    color: #f56c6c;
-  }
-
-  .success {
-    color: #67c23a;
-  }
-
-  .warning {
-    color: #e6a23c;
-  }
-
-  .connections-status__icon {
-    font-size: 14px;
-  }
-
-  .search-bar {
-    display: flex;
-
-    .item {
-      margin-right: 10px;
-    }
-  }
-
-  .btn + .btn {
+  &.btn-dropdowm {
     margin-left: 5px;
   }
 
-  .btn {
-    i.iconfont {
-      font-size: 12px;
-    }
-
-    &.btn-dropdowm {
-      margin-left: 5px;
-    }
-
-    &.btn-create {
-      margin-left: 5px;
-    }
+  &.btn-create {
+    margin-left: 5px;
   }
 }
 </style>
