@@ -140,7 +140,7 @@ export default {
   },
   data() {
     return {
-      isDaas: import.meta.env.VITE_PLATFORM === 'DAAS',
+      isDaas: import.meta.env.VUE_APP_PLATFORM === 'DAAS',
       drawer: false,
       visible: false,
       timer: null,
