@@ -1,71 +1,85 @@
 <template>
-  <section class="data-server-wrapper flex flex-column">
-    <div v-if="showFilter" class="flex justify-content-between my-2">
-      <FilterBar v-model:value="searchParams" :items="filterItems" @fetch="table.fetch(1)"> </FilterBar>
-      <div>
-        <ElButton
-          v-show="multipleSelection.length > 0"
-          :disabled="$disabledReadonlyUserBtn()"
-          v-readonlybtn="'SYNC_job_export'"
-          class="btn message-button-cancel"
-          @click="handleExport"
-        >
-          <span> {{ $t('public_button_export') }}</span>
-        </ElButton>
-        <ElButton
-          v-readonlybtn="'SYNC_job_import'"
-          class="btn"
-          :disabled="$disabledReadonlyUserBtn()"
-          @click="handleImport"
-        >
-          <span> {{ $t('packages_business_button_bulk_import') }}</span>
-        </ElButton>
-        <ElButton
-          v-readonlybtn="'SYNC_job_export'"
-          class="btn"
-          :disabled="$disabledReadonlyUserBtn() || !multipleSelectionActive.length"
-          @click="handleExportApiDoc"
-        >
-          <span>{{ $t('packages_business_data_server_list_apIwendang') }}</span>
-        </ElButton>
-        <ElButton class="btn btn-create" type="primary" @click.stop="showDrawer()">
-          <span>{{ $t('packages_business_data_server_drawer_chuangjianfuwu') }}</span>
-        </ElButton>
-      </div>
-    </div>
-    <VTable
-      :columns="cols"
-      :remote-method="getData"
-      ref="table"
-      height="100%"
-      class="flex-fill"
-      @selection-change="handleSelectionChange"
-    >
-      <template #name="{ row }">
-        <ElLink class="ellipsis" type="primary" style="display: block; line-height: 20px" @click.stop="showDrawer(row)">
-          {{ row.name }}
-        </ElLink>
-      </template>
-      <template #statusFmt="{ row }">
-        <span class="status-block" :class="'status-' + row.status">{{ row.statusFmt }}</span>
-      </template>
-      <template #operation="{ row }">
-        <ElButton v-if="row.status !== 'active'" :disabled="row.status !== 'pending'" text @click="changeStatus(row)">{{
-          $t('public_button_public')
-        }}</ElButton>
-        <ElButton v-if="row.status === 'active'" text @click="changeStatus(row)">{{
-          $t('public_button_revoke')
-        }}</ElButton>
-        <ElDivider direction="vertical"></ElDivider>
-        <ElButton text @click="output(row)">{{ $t('public_button_export') }}</ElButton>
-        <ElDivider direction="vertical"></ElDivider>
-        <ElButton text @click="removeServer(row)">{{ $t('public_button_delete') }}</ElButton>
-      </template>
+  <PageContainer>
+    <template #actions>
+      <ElButton
+        v-show="multipleSelection.length > 0"
+        :disabled="$disabledReadonlyUserBtn()"
+        v-readonlybtn="'SYNC_job_export'"
+        class="btn message-button-cancel"
+        @click="handleExport"
+      >
+        <span> {{ $t('public_button_export') }}</span>
+      </ElButton>
+      <ElButton
+        v-readonlybtn="'SYNC_job_import'"
+        class="btn"
+        :disabled="$disabledReadonlyUserBtn()"
+        @click="handleImport"
+      >
+        <span> {{ $t('packages_business_button_bulk_import') }}</span>
+      </ElButton>
+      <ElButton
+        v-readonlybtn="'SYNC_job_export'"
+        class="btn"
+        :disabled="$disabledReadonlyUserBtn() || !multipleSelectionActive.length"
+        @click="handleExportApiDoc"
+      >
+        <span>{{ $t('packages_business_data_server_list_apIwendang') }}</span>
+      </ElButton>
+      <ElButton class="btn btn-create" type="primary" @click.stop="showDrawer()">
+        <span>{{ $t('packages_business_data_server_drawer_chuangjianfuwu') }}</span>
+      </ElButton>
+    </template>
 
-      <template v-slot:empty>
-        <VEmpty large></VEmpty>
-      </template>
-    </VTable>
+    <div class="flex flex-column h-100">
+      <div v-if="showFilter" class="flex justify-content-between p-3">
+        <FilterBar v-model:value="searchParams" :items="filterItems" @fetch="table.fetch(1)"> </FilterBar>
+      </div>
+
+      <VTable
+        :columns="cols"
+        :remote-method="getData"
+        ref="table"
+        height="100%"
+        class="flex-fill"
+        @selection-change="handleSelectionChange"
+      >
+        <template #name="{ row }">
+          <ElLink
+            class="ellipsis"
+            type="primary"
+            style="display: block; line-height: 20px"
+            @click.stop="showDrawer(row)"
+          >
+            {{ row.name }}
+          </ElLink>
+        </template>
+        <template #statusFmt="{ row }">
+          <span class="status-block" :class="'status-' + row.status">{{ row.statusFmt }}</span>
+        </template>
+        <template #operation="{ row }">
+          <ElButton
+            v-if="row.status !== 'active'"
+            :disabled="row.status !== 'pending'"
+            text
+            @click="changeStatus(row)"
+            >{{ $t('public_button_public') }}</ElButton
+          >
+          <ElButton v-if="row.status === 'active'" text @click="changeStatus(row)">{{
+            $t('public_button_revoke')
+          }}</ElButton>
+          <ElDivider direction="vertical"></ElDivider>
+          <ElButton text @click="output(row)">{{ $t('public_button_export') }}</ElButton>
+          <ElDivider direction="vertical"></ElDivider>
+          <ElButton text @click="removeServer(row)">{{ $t('public_button_delete') }}</ElButton>
+        </template>
+
+        <template v-slot:empty>
+          <VEmpty large></VEmpty>
+        </template>
+      </VTable>
+    </div>
+
     <Drawer
       ref="drawer"
       :host="apiServerHost"
@@ -74,7 +88,7 @@
     ></Drawer>
     <!-- 导入 -->
     <Upload type="Modules" :show-tag="false" ref="upload" @success="table.fetch()"></Upload>
-  </section>
+  </PageContainer>
 </template>
 
 <script>
@@ -86,9 +100,10 @@ import { FilterBar, VTable, VEmpty } from '@tap/component'
 import Upload from '../../components/UploadDialog'
 
 import Drawer from './Drawer'
+import PageContainer from '../../components/PageContainer.vue'
 
 export default {
-  components: { FilterBar, Drawer, VTable, VEmpty, Upload },
+  components: { PageContainer, FilterBar, Drawer, VTable, VEmpty, Upload },
   props: {
     showFilter: {
       type: Boolean,
@@ -177,7 +192,7 @@ export default {
         {
           label: this.$t('daas_data_server_drawer_path'),
           'min-width': 130,
-          prop: '_path'
+          prop: '_path',
         },
         {
           label: this.$t('packages_business_data_server_list_fuwuzhuangtai'),
