@@ -21,197 +21,191 @@
       </ElButton>
     </template>
 
-    <section class="data-verify-wrap bg-white section-wrap">
-      <TablePage
-        ref="table"
-        row-key="id"
-        :remoteMethod="getData"
-        @selection-change="handleSelectionChange"
-        @sort-change="handleSortTable"
+    <TablePage
+      ref="table"
+      row-key="id"
+      :remoteMethod="getData"
+      @selection-change="handleSelectionChange"
+      @sort-change="handleSortTable"
+    >
+      <template #search>
+        <div class="search-bar">
+          <FilterBar v-model:value="searchParams" :items="filterItems" @fetch="table.fetch(1)"></FilterBar>
+        </div>
+      </template>
+
+      <template #multipleSelectionActions>
+        <ElButton @click="handlePermissionsSettings">{{
+          $t('packages_business_permissionse_settings_create_quanxianshezhi')
+        }}</ElButton>
+        <ElButton v-readonlybtn="'SYNC_category_application'" size="mini" class="btn" @click="handleExport">
+          <i class="iconfont icon-daoru back-btn-icon"></i>
+          <span> {{ $t('public_button_export') }}</span>
+        </ElButton>
+      </template>
+
+      <el-table-column reserve-selection type="selection" width="38" align="center"></el-table-column>
+      <el-table-column :label="$t('packages_business_verification_task_name')" min-width="250" show-overflow-tooltip>
+        <template v-slot="scope">
+          <div class="ellipsis">{{ scope.row.name }}</div>
+          <div class="font-color-slight">
+            <span
+              >{{ getInspectName(scope.row) }} (
+              {{
+                scope.row.mode === 'manual'
+                  ? $t('packages_business_verification_singleVerify')
+                  : $t('packages_business_verification_repeatingVerify')
+              }}
+              )
+            </span>
+            <span v-if="!scope.row.enabled" class="font-color-slight">&nbsp;Disabled</span>
+          </div>
+        </template>
+      </el-table-column>
+      <el-table-column
+        prop="sourceTotal"
+        min-width="140"
+        align="center"
+        :label="$t('packages_business_verification_history_source_total_rows')"
       >
-        <template #search>
-          <div class="search-bar">
-            <FilterBar v-model="searchParams" :items="filterItems" @fetch="table.fetch(1)"></FilterBar>
+        <template v-slot="scope">
+          {{ scope.row.inspectMethod === 'hash' ? '-' : scope.row.sourceTotal || 0 }}
+        </template>
+      </el-table-column>
+      <el-table-column :label="$t('packages_business_verification_result_title')" min-width="180">
+        <template v-slot="scope">
+          <div class="flex align-center">
+            <template v-if="scope.row.InspectResult && ['waiting', 'done'].includes(scope.row.status)">
+              <div v-if="scope.row.result !== 'passed'" class="data-verify__status error">
+                <i class="data-verify__icon el-icon-error"></i>
+                <span v-if="scope.row.inspectMethod === 'row_count' || scope.row.inspectMethod === 'hash'">
+                  {{ $t('packages_business_verification_inconsistent') }}
+                </span>
+                <span v-if="scope.row.inspectMethod === 'field'">
+                  {{ $t('packages_business_verification_contConsistent') }} {{ scope.row.difference_number }}
+                </span>
+                <span v-if="scope.row.inspectMethod === 'jointField'">
+                  {{ $t('packages_business_verification_contConsistent') }} {{ scope.row.difference_number }}
+                </span>
+                <span v-if="scope.row.inspectMethod === 'cdcCount'">
+                  {{ $t('packages_business_verification_contConsistent') }} {{ scope.row.difference_number }}
+                </span>
+              </div>
+              <div v-else class="data-verify__status success">
+                <i class="data-verify__icon el-icon-success"></i>
+                <span>{{ $t('packages_business_verification_consistent') }}</span>
+              </div>
+            </template>
+            <div v-else-if="scope.row.status === 'error'" class="data-verify__status">
+              <i class="data-verify__icon el-icon-error"></i>
+              <span>{{ $t('public_status_error') }}</span>
+              <ElLink type="primary" class="ml-2" @click="handleError(scope.row)"
+                >{{ $t('public_button_check') }}
+              </ElLink>
+            </div>
+            <div v-else-if="scope.row.status === 'waiting'" class="data-verify__status">-</div>
+            <div v-else-if="scope.row.status !== 'done'" class="data-verify__status">
+              <img style="width: 26px; vertical-align: middle" :src="loadingImg" />
+              <span>{{ statusMap[scope.row.status] }}</span>
+            </div>
+            <div v-else>-</div>
+            <!--            <VIcon v-if="scope.row.InspectResult && scope.row.InspectResult.parentId" class="ml-2" size="16"-->
+            <!--              >ercijiaoyan</VIcon-->
+            <!--            >-->
+          </div>
+        </template>
+      </el-table-column>
+      <el-table-column :label="$t('packages_business_verification_verifyStatus')" min-width="110" prop="status">
+        <template v-slot="scope">
+          <span>{{ statusMap[scope.row.status] }}</span>
+          <span v-if="scope.row.InspectResult && scope.row.status === 'running'">
+            {{ `(${scope.row.InspectResult.progress ? Math.floor(scope.row.InspectResult.progress * 100) : 0}%)` }}
+          </span>
+        </template>
+      </el-table-column>
+      <el-table-column
+        :label="$t('packages_business_verification_verifyTime')"
+        prop="lastStartTime"
+        sortable="lastStartTime"
+        min-width="170"
+      ></el-table-column>
+      <el-table-column :label="$t('public_operation')" width="260">
+        <template v-if="isDaas" #header>
+          <div class="flex align-center">
+            <span>{{ $t('public_operation_available') }}</span>
+            <ElTooltip class="ml-2" placement="top" :content="$t('packages_business_connections_list_wuquanxiandecao')">
+              <VIcon class="color-primary" size="14">info</VIcon>
+            </ElTooltip>
           </div>
         </template>
 
-        <template #multipleSelectionActions>
-          <ElButton @click="handlePermissionsSettings">{{
-            $t('packages_business_permissionse_settings_create_quanxianshezhi')
-          }}</ElButton>
-          <ElButton v-readonlybtn="'SYNC_category_application'" size="mini" class="btn" @click="handleExport">
-            <i class="iconfont icon-daoru back-btn-icon"></i>
-            <span> {{ $t('public_button_export') }}</span>
-          </ElButton>
-        </template>
+        <template #default="{ row }">
+          <ElLink type="primary" :disabled="!row.InspectResult" @click="toTableInfo(row.id)"
+            >{{ $t('packages_business_verification_result_title') }}
+          </ElLink>
 
-        <el-table-column reserve-selection type="selection" width="38" align="center"></el-table-column>
-        <el-table-column :label="$t('packages_business_verification_task_name')" min-width="250" show-overflow-tooltip>
-          <template slot-scope="scope">
-            <div class="ellipsis">{{ scope.row.name }}</div>
-            <div class="font-color-slight">
-              <span
-                >{{ getInspectName(scope.row) }} (
-                {{
-                  scope.row.mode === 'manual'
-                    ? $t('packages_business_verification_singleVerify')
-                    : $t('packages_business_verification_repeatingVerify')
-                }}
-                )
-              </span>
-              <span v-if="!scope.row.enabled" class="font-color-slight">&nbsp;Disabled</span>
-            </div>
-          </template>
-        </el-table-column>
-        <el-table-column
-          prop="sourceTotal"
-          min-width="140"
-          align="center"
-          :label="$t('packages_business_verification_history_source_total_rows')"
-        >
-          <template slot-scope="scope">
-            {{ scope.row.inspectMethod === 'hash' ? '-' : scope.row.sourceTotal || 0 }}
-          </template>
-        </el-table-column>
-        <el-table-column :label="$t('packages_business_verification_result_title')" min-width="180">
-          <template slot-scope="scope">
-            <div class="flex align-center">
-              <template v-if="scope.row.InspectResult && ['waiting', 'done'].includes(scope.row.status)">
-                <div v-if="scope.row.result !== 'passed'" class="data-verify__status error">
-                  <i class="data-verify__icon el-icon-error"></i>
-                  <span v-if="scope.row.inspectMethod === 'row_count' || scope.row.inspectMethod === 'hash'">
-                    {{ $t('packages_business_verification_inconsistent') }}
-                  </span>
-                  <span v-if="scope.row.inspectMethod === 'field'">
-                    {{ $t('packages_business_verification_contConsistent') }} {{ scope.row.difference_number }}
-                  </span>
-                  <span v-if="scope.row.inspectMethod === 'jointField'">
-                    {{ $t('packages_business_verification_contConsistent') }} {{ scope.row.difference_number }}
-                  </span>
-                  <span v-if="scope.row.inspectMethod === 'cdcCount'">
-                    {{ $t('packages_business_verification_contConsistent') }} {{ scope.row.difference_number }}
-                  </span>
-                </div>
-                <div v-else class="data-verify__status success">
-                  <i class="data-verify__icon el-icon-success"></i>
-                  <span>{{ $t('packages_business_verification_consistent') }}</span>
-                </div>
-              </template>
-              <div v-else-if="scope.row.status === 'error'" class="data-verify__status">
-                <i class="data-verify__icon el-icon-error"></i>
-                <span>{{ $t('public_status_error') }}</span>
-                <ElLink type="primary" class="ml-2" @click="handleError(scope.row)"
-                  >{{ $t('public_button_check') }}
-                </ElLink>
-              </div>
-              <div v-else-if="scope.row.status === 'waiting'" class="data-verify__status">-</div>
-              <div v-else-if="scope.row.status !== 'done'" class="data-verify__status">
-                <img style="width: 26px; vertical-align: middle" :src="loadingImg" />
-                <span>{{ statusMap[scope.row.status] }}</span>
-              </div>
-              <div v-else>-</div>
-              <!--            <VIcon v-if="scope.row.InspectResult && scope.row.InspectResult.parentId" class="ml-2" size="16"-->
-              <!--              >ercijiaoyan</VIcon-->
-              <!--            >-->
-            </div>
-          </template>
-        </el-table-column>
-        <el-table-column :label="$t('packages_business_verification_verifyStatus')" min-width="110" prop="status">
-          <template slot-scope="scope">
-            <span>{{ statusMap[scope.row.status] }}</span>
-            <span v-if="scope.row.InspectResult && scope.row.status === 'running'">
-              {{ `(${scope.row.InspectResult.progress ? Math.floor(scope.row.InspectResult.progress * 100) : 0}%)` }}
-            </span>
-          </template>
-        </el-table-column>
-        <el-table-column
-          :label="$t('packages_business_verification_verifyTime')"
-          prop="lastStartTime"
-          sortable="lastStartTime"
-          min-width="170"
-        ></el-table-column>
-        <el-table-column :label="$t('public_operation')" width="260">
-          <template v-if="isDaas" #header>
-            <div class="flex align-center">
-              <span>{{ $t('public_operation_available') }}</span>
-              <ElTooltip
-                class="ml-2"
-                placement="top"
-                :content="$t('packages_business_connections_list_wuquanxiandecao')"
-              >
-                <VIcon class="color-primary" size="14">info</VIcon>
-              </ElTooltip>
-            </div>
-          </template>
-
-          <template #default="{ row }">
-            <ElLink type="primary" :disabled="!row.InspectResult" @click="toTableInfo(row.id)"
-              >{{ $t('packages_business_verification_result_title') }}
+          <template v-if="havePermission(row.permissionActions, 'Stop') && row.status === 'running'">
+            <ElDivider direction="vertical" v-readonlybtn="'verify_job_edition'"></ElDivider>
+            <ElLink
+              v-readonlybtn="'verify_job_edition'"
+              type="primary"
+              :disabled="$disabledByPermission('verify_job_edition_all_data', row.user_id)"
+              @click="stop(row.id)"
+              >{{ $t('public_button_stop') }}
             </ElLink>
+          </template>
 
-            <template v-if="havePermission(row.permissionActions, 'Stop') && row.status === 'running'">
-              <ElDivider direction="vertical" v-readonlybtn="'verify_job_edition'"></ElDivider>
-              <ElLink
-                v-readonlybtn="'verify_job_edition'"
-                type="primary"
-                :disabled="$disabledByPermission('verify_job_edition_all_data', row.user_id)"
-                @click="stop(row.id)"
-                >{{ $t('public_button_stop') }}
-              </ElLink>
-            </template>
-
-            <template v-if="havePermission(row.permissionActions, 'Start') && row.status !== 'running'">
-              <ElDivider direction="vertical"></ElDivider>
-              <ElLink
-                v-readonlybtn="'verify_job_edition'"
-                type="primary"
-                :disabled="
-                  $disabledByPermission('verify_job_edition_all_data', row.user_id) ||
-                  ['running', 'scheduling', 'stopping'].includes(row.status)
-                "
-                @click="startTask(row.id)"
-                >{{ $t('packages_business_verification_executeVerifyTip') }}
-              </ElLink>
-            </template>
-
+          <template v-if="havePermission(row.permissionActions, 'Start') && row.status !== 'running'">
             <ElDivider direction="vertical"></ElDivider>
             <ElLink
               v-readonlybtn="'verify_job_edition'"
               type="primary"
-              :disabled="!row.InspectResult"
-              @click="history(row.id)"
-              >{{ $t('packages_business_verification_historyTip') }}
+              :disabled="
+                $disabledByPermission('verify_job_edition_all_data', row.user_id) ||
+                ['running', 'scheduling', 'stopping'].includes(row.status)
+              "
+              @click="startTask(row.id)"
+              >{{ $t('packages_business_verification_executeVerifyTip') }}
             </ElLink>
-
-            <template v-if="havePermission(row.permissionActions, 'Edit')">
-              <ElDivider direction="vertical" v-readonlybtn="'verify_job_edition'"></ElDivider>
-              <ElLink
-                v-readonlybtn="'verify_job_edition'"
-                type="primary"
-                :disabled="
-                  $disabledByPermission('verify_job_edition_all_data', row.user_id) ||
-                  ['running', 'scheduling'].includes(row.status)
-                "
-                @click="goEdit(row.id, row.flowId)"
-                >{{ $t('packages_business_verification_configurationTip') }}
-              </ElLink>
-            </template>
-
-            <template v-if="havePermission(row.permissionActions, 'Delete')">
-              <ElDivider direction="vertical"></ElDivider>
-              <ElLink
-                v-readonlybtn="'verify_job_edition'"
-                type="primary"
-                :disabled="$disabledByPermission('verify_job_delete_all_data', row.user_id)"
-                @click="remove(row.id, row)"
-                >{{ $t('public_button_delete') }}
-              </ElLink>
-            </template>
           </template>
-        </el-table-column>
-      </TablePage>
-      <PermissionseSettingsCreate ref="permissionseSettingsCreate"></PermissionseSettingsCreate>
-    </section>
+
+          <ElDivider direction="vertical"></ElDivider>
+          <ElLink
+            v-readonlybtn="'verify_job_edition'"
+            type="primary"
+            :disabled="!row.InspectResult"
+            @click="history(row.id)"
+            >{{ $t('packages_business_verification_historyTip') }}
+          </ElLink>
+
+          <template v-if="havePermission(row.permissionActions, 'Edit')">
+            <ElDivider direction="vertical" v-readonlybtn="'verify_job_edition'"></ElDivider>
+            <ElLink
+              v-readonlybtn="'verify_job_edition'"
+              type="primary"
+              :disabled="
+                $disabledByPermission('verify_job_edition_all_data', row.user_id) ||
+                ['running', 'scheduling'].includes(row.status)
+              "
+              @click="goEdit(row.id, row.flowId)"
+              >{{ $t('packages_business_verification_configurationTip') }}
+            </ElLink>
+          </template>
+
+          <template v-if="havePermission(row.permissionActions, 'Delete')">
+            <ElDivider direction="vertical"></ElDivider>
+            <ElLink
+              v-readonlybtn="'verify_job_edition'"
+              type="primary"
+              :disabled="$disabledByPermission('verify_job_delete_all_data', row.user_id)"
+              @click="remove(row.id, row)"
+              >{{ $t('public_button_delete') }}
+            </ElLink>
+          </template>
+        </template>
+      </el-table-column>
+    </TablePage>
+    <PermissionseSettingsCreate ref="permissionseSettingsCreate"></PermissionseSettingsCreate>
   </PageContainer>
 </template>
 
@@ -227,6 +221,7 @@ import { inspectApi, metadataInstancesApi } from '@tap/api'
 import { statusMap, inspectMethod, typeList as verifyTypeList } from './const'
 import PageContainer from '../../components/PageContainer.vue'
 import PermissionseSettingsCreate from '../../components/permissionse-settings/Create'
+import loadingImg from '@tap/assets/icons/loading.svg'
 
 let timeout = null
 export default {
@@ -250,7 +245,7 @@ export default {
         result: '',
       },
       filterItems: [],
-      loadingImg: require('@tap/assets/icons/loading.svg'),
+      loadingImg,
       order: 'last_updated DESC',
       inspectMethod,
       statusMap,
@@ -554,8 +549,8 @@ export default {
     },
     handlePermissionsSettings() {
       this.$refs.permissionseSettingsCreate.open(this.multipleSelection, 'Inspect')
-    }
-  }
+    },
+  },
 }
 </script>
 
