@@ -78,7 +78,7 @@ export default {
 
   data() {
     return {
-      isDaas: import.meta.env.VITE_PLATFORM === 'DAAS',
+      isDaas: import.meta.env.VUE_APP_PLATFORM === 'DAAS',
       currentTab: 'settings',
       titleCurrentTab: 'settings',
       name: this.activeNode?.name,
