@@ -136,7 +136,7 @@ export default {
     return {
       keyword: '',
       visible: false,
-      isDaas: import.meta.env.VITE_PLATFORM === 'DAAS',
+      isDaas: import.meta.env.VUE_APP_PLATFORM === 'DAAS',
       showSceneDialog: false,
       settingsVisible: false,
       dragState: {
@@ -152,7 +152,7 @@ export default {
         fdmStorageCluster: 'self',
         fdmStorageConnectionId: '',
         mdmStorageCluster: 'self',
-        mdmStorageConnectionId: ''
+        mdmStorageConnectionId: '',
       },
       directoryMap: {},
       fdmConnection: null,
@@ -425,14 +425,14 @@ export default {
             sourceNode.type === 'apiserverLineage'
               ? 'target'
               : this.settings.fdmStorageConnectionId === sourceNode.connectionId
-              ? otherLdpType
-              : 'source'
+                ? otherLdpType
+                : 'source'
           targetNode.ldpType =
             targetNode.type === 'apiserverLineage'
               ? 'target'
               : this.settings.fdmStorageConnectionId === targetNode.connectionId
-              ? otherLdpType
-              : 'source'
+                ? otherLdpType
+                : 'source'
           // 记录事件触发的dom和ldpType
           if (sourceNode.table === tableInfo.name) {
             sourceNode.ldpType = ldpType
@@ -642,8 +642,8 @@ export default {
           this.jsPlumbIns.revalidate(el)
         }
       }
-    }
-  }
+    },
+  },
 }
 </script>
 
